componentize a few careers page sections
