+++ docs/src/modules/components/TopLayoutCareers.js
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import { styled } from '@mui/material/styles';
+import { styled, alpha } from '@mui/material/styles';
 import Divider from '@mui/material/Divider';
+import KeyboardArrowLeftIcon from '@mui/icons-material/KeyboardArrowLeft';
 import Head from 'docs/src/modules/components/Head';
 import AppContainer from 'docs/src/modules/components/AppContainer';
 import AppFooter from 'docs/src/layouts/AppFooter';
@@ -10,9 +11,17 @@ import BrandingCssVarsProvider from 'docs/src/BrandingCssVarsProvider';
 import MarkdownElement from 'docs/src/modules/components/MarkdownElement';
 import { Link } from '@mui/docs/Link';
 
-const StyledDiv = styled('div')({
+const StyledDiv = styled('div')(({ theme }) => ({
   flex: '1 0 100%',
-});
+  background: `linear-gradient(180deg, ${(theme.vars || theme).palette.grey[50]} 0%, #FFFFFF 100%)`,
+  backgroundSize: '100% 500px',
+  backgroundRepeat: 'no-repeat',
+  ...theme.applyDarkStyles({
+    background: `linear-gradient(180deg, ${alpha(theme.palette.primary[900], 0.15)} 0%, ${(theme.vars || theme).palette.primaryDark[900]} 100%)`,
+    backgroundSize: '100% 500px',
+    backgroundRepeat: 'no-repeat',
+  }),
+}));
 
 const StyledAppContainer = styled(AppContainer)(({ theme }) => ({
   '& .markdownElement': {
@@ -42,10 +51,11 @@ export default function TopLayoutCareers(props) {
             href="/careers/#open-roles"
             rel="nofollow"
             variant="body2"
-            sx={{ display: 'block', mb: 2 }}
+            sx={{ display: 'flex', gap: 0.5, alignItems: 'center', mb: 2 }}
           >
+            <KeyboardArrowLeftIcon fontSize="small" />
             {/* eslint-disable-next-line material-ui/no-hardcoded-labels */}
-            {'< Back to open roles'}
+            {'Back to open roles'}
           </Link>
           {rendered.map((chunk, index) => {
             return <MarkdownElement key={index} renderedMarkdown={chunk} />;
